Bug Fix.
There was a bug found when adding a new build and you tried to add commands to it. The dropdown was not getting populated because of the NOT IN query, which was getting an empty array and therefore not returning any results to populate the dropdown. Added extra checks to fix this. so when there are no build commands we return all commands.